Refactored and moved methods to the OpenRouterServcies.ts and updated providers.ts to match the rest of the providers.
Also updated the referer URL.
